Add files via upload
This is my BootCamp coding project. Whilst coding this piece of animation I was able to test my coding skills at a much higher level. In the past, I have coded using C# which is somewhat similar to Java. Coding with Java showed my different ways to create variables for different methods and functions which is what I practiced a lot through this exercise of mine. That is one big difference I saw between C# and Java. Being able to do this in Java has now made me much more comfortable in using the language as a whole.